added test image handler, minio support, dtos for GET

+++ Application/Services/ProjectService.cs
@@ -2,48 +2,43 @@
 using Constructor_API.Core.Repositories;
 using Constructor_API.Helpers.Exceptions;
 using Constructor_API.Models.DTOs.Create;
+using Constructor_API.Models.DTOs.Read;
 using Constructor_API.Models.DTOs.Update;
 using Constructor_API.Models.Entities;
-using Constructor_API.Models.Objects;
-using Microsoft.AspNetCore.Authorization;
 using MongoDB.Bson;
 using System.Linq;
-using System.Threading;
-using System.Xml.Linq;
 
 namespace Constructor_API.Application.Services
 {
     public class ProjectService
     {
         IProjectRepository _projectRepository;
-        //IProjectUserRepository _projectUserRepository;
         IBuildingRepository _buildingRepository;
-        //IFloorRepository _floorRepository;
-        //IGraphPointRepository _graphPointRepository;
-        //IFloorConnectionRepository _floorConnectionRepository;
         IUserRepository _userRepository;
         IMapper _mapper;
+        IProjectUserRepository _projectUserRepository;
 
-        public ProjectService(IProjectRepository projectRepository, IBuildingRepository buildingRepository,
-            IUserRepository userRepository, /*IFloorRepository floorRepository, 
-            IGraphPointRepository floorGraphPointRepository, IFloorConnectionRepository floorConnectionRepository,*/
-            IMapper mapper/*, IProjectUserRepository projectUserRepository*/)
+        public ProjectService(
+            IProjectRepository projectRepository, 
+            IBuildingRepository buildingRepository,
+            IUserRepository userRepository,
+            IMapper mapper,
+            IProjectUserRepository projectUserRepository)
         {
             _projectRepository = projectRepository;
             _buildingRepository = buildingRepository;
             _userRepository = userRepository;
-            //_floorRepository = floorRepository;
-            //_graphPointRepository = floorGraphPointRepository;
-            //_floorConnectionRepository = floorConnectionRepository;
             _mapper = mapper;
-            //_projectUserRepository = projectUserRepository;
+            _projectUserRepository = projectUserRepository;
         }
 
-        public async Task InsertProject(
-            CreateProjectDto projectDto, string userId, CancellationToken cancellationToken)
+        public async Task<Project> InsertProject(
+            CreateProjectDto projectDto, 
+            string userId, 
+            CancellationToken cancellationToken)
         {
-            var user = await _userRepository.FirstOrDefaultAsync(u => u.Id == userId, cancellationToken)
-                ?? throw new NotFoundException("User is not found");
+            if (await _userRepository.CountAsync(u => u.Id == userId, cancellationToken) == 0)
+                throw new NotFoundException("User is not found");
 
             Project project = _mapper.Map<Project>(projectDto);
             project.CustomGraphPointTypes = [];
@@ -62,76 +57,55 @@ namespace Constructor_API.Application.Services
                 AddedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
             };
-            ///
             project.ProjectUsers = [projUser];
 
             //await _projectUserRepository.AddAsync(projUser, cancellationToken);
             await _projectRepository.AddAsync(project, cancellationToken);
             await _projectRepository.SaveChanges();
+
+            return project;
         }
 
         public async Task<Project> GetProjectById(
             string id, CancellationToken cancellationToken)
         {
-            var res = await _projectRepository.FirstOrDefaultAsync(g => g.Id == id, cancellationToken);
-            if (res == null) throw new NotFoundException("Navigation group not found");
+            return await _projectRepository.FirstOrDefaultAsync(g => g.Id == id, cancellationToken)
+                ?? throw new NotFoundException("Project is not found");
+        }
 
-            return res;
+        public async Task<GetProjectDto> GetProjectInfoById(
+            string id, CancellationToken cancellationToken)
+        {
+            return await _projectRepository.FirstGetProjectDtoOrDefaultAsync(g => g.Id == id, cancellationToken)
+                ?? throw new NotFoundException("Project is not found");
         }
 
         public async Task<IReadOnlyList<Project>> GetAllProjects(
             CancellationToken cancellationToken)
         {
-            var res = await _projectRepository.ListAsync(cancellationToken);
-            return res;
+            return await _projectRepository.ListAsync(cancellationToken);
         }
 
-        public async Task<IReadOnlyList<Building>> GetBuildingsByProject(string projectId,
+        public async Task<IReadOnlyList<GetBuildingDto>> GetBuildingsByProject(string projectId,
             CancellationToken cancellationToken)
         {
-            var res = await _buildingRepository.ListAsync(b => b.ProjectId == projectId, cancellationToken);
-            return res;
+            return await _buildingRepository.SimpleGetBuildingDtoListAsync(b => b.ProjectId == projectId, cancellationToken);
         }
 
         public async Task<Building> GetBuildingInProjectByName(string projectId, string name,
             CancellationToken cancellationToken)
         {
-            var building = await _buildingRepository.FirstOrDefaultAsync(b => b.ProjectId == projectId && b.Name == name,
-                cancellationToken);
-            if (building == null) throw new NotFoundException("Building not found");
-
-            return building;
+            return await _buildingRepository.FirstOrDefaultAsync(b =>
+                b.ProjectId == projectId && b.Name == name, cancellationToken)
+                ?? throw new NotFoundException("Building not found");
         }
 
         public async Task DeleteProject(string id, CancellationToken cancellationToken)
         {
-            var project = await _projectRepository.FirstOrDefaultAsync(p => p.Id == id, cancellationToken)
-                ?? throw new NotFoundException($"Project is not found");
-            var buildings = await _buildingRepository.ListAsync(b => b.ProjectId == id, cancellationToken)
-                ?? throw new NotFoundException($"Buildings are not found");
-            ////var user = await _userRepository.FirstOrDefaultAsync(u => u.Id == userId, cancellationToken)
-            ////    ?? throw new NotFoundException("User is not found");
-            ////user.ProjectIds = user.ProjectIds.Where(p => p != id).ToArray();
-            ////user.UpdatedAt = DateTime.UtcNow;
-
-            //await _floorConnectionRepository.RemoveRangeAsync(c => buildings.Any(b => b.Id == c.BuildingId), cancellationToken);
-
-            //foreach (var building in buildings)
-            //{
-            //    if (building.FloorIds != null)
-            //    {
-            //        foreach (var floorId in building.FloorIds)
-            //        {
-            //            var floor = await _floorRepository.FirstOrDefaultAsync(b => b.Id == floorId, cancellationToken)
-            //                ?? throw new NotFoundException("Floor is not found");
-
-            //            await _graphPointRepository.RemoveRangeAsync(g => floor.GraphPoints.Contains(g.Id), cancellationToken);
-            //        }
-            //        await _floorRepository.RemoveRangeAsync(f => f.BuildingId == building.Id, cancellationToken);
-            //    }
-            //}
-
-            //await _buildingRepository.RemoveRangeAsync(b => b.ProjectId == id, cancellationToken);
+            if (await _projectRepository.CountAsync(p => p.Id == id, cancellationToken) == 0)
+                throw new NotFoundException($"Project is not found");
+            if (await _buildingRepository.CountAsync(b => b.ProjectId == id, cancellationToken) == 0)
+                throw new NotFoundException($"Buildings are not found");
             await _projectRepository.RemoveAsync(p => p.Id == id, cancellationToken);
             await _projectRepository.SaveChanges();
         }
@@ -144,7 +118,7 @@ namespace Constructor_API.Application.Services
             prevProject.Name = projectDto.Name ?? prevProject.Name;
             prevProject.Url = projectDto.Url ?? prevProject.Url;
             prevProject.Description = projectDto.Description ?? prevProject.Description;
-            prevProject.ImageId = projectDto.ImageId ?? prevProject.ImageId;
+            //prevProject.ImageId = projectDto.ImageId ?? prevProject.ImageId;
             prevProject.CustomGraphPointTypes = projectDto.CustomGraphPointTypes ?? prevProject.CustomGraphPointTypes;
             prevProject.UpdatedAt = DateTime.UtcNow;
